remaining types of program counters changed to gsize/gssize
* This fixes F< to the beginning of the macro, which was broken in 73d574b71a10d4661ada20275cafde75aff6c1ba.
  teco_machine_main_t::macro_pc actually has to be signed as it is sometimes set to -1.

+++ src/cmdline.c
@@ -87,7 +87,7 @@ teco_cmdline_insert(const gchar *data, gsize len, GError **error)
 {
 	const teco_string_t src = {(gchar *)data, len};
 	teco_string_t old_cmdline = {NULL, 0};
-	guint repl_pc = 0;
+	gsize repl_pc = 0;
 
 	teco_cmdline.machine.macro_pc = teco_cmdline.pc = teco_cmdline.effective_len;
